refactor(users): extract legacy page and add feature flag routing
