+++ app/admin/users/users-v2.tsx
@@ -1,0 +1,5 @@
+'use client';
+
+export default function UsersV2() {
+  return <div>Users V2 - placeholder</div>;
+}
